feat(house): Hue light setup and control through route API (#173)
