+++ src/routes/api/control/hue/[collection]/[id]/[updateType]/+server.ts
@@ -1,0 +1,31 @@
+import { json } from '@sveltejs/kit';
+import {
+  SECRET_HUE_IP_ADDRESS,
+  SECRET_HUE_USERNAME,
+} from '$env/static/private';
+
+import type { RequestHandler } from './$types';
+import type { FetchOptions } from '@types';
+
+export const PUT: RequestHandler = async ({
+  fetch,
+  request,
+  params: { collection, id, updateType },
+}) => {
+  const { on } = await request.json();
+  const headers = new Headers();
+  headers.append('Content-Type', 'application/json');
+  const requestOptions: FetchOptions = {
+    method: 'PUT',
+    redirect: 'follow',
+    body: JSON.stringify({
+      on,
+    }),
+  };
+  const requestURL = `https://${SECRET_HUE_IP_ADDRESS}/api/${SECRET_HUE_USERNAME}/${collection}${id ? `/${id}` : ''}${updateType ? `/${updateType}` : ''}`;
+  await fetch(requestURL, requestOptions)
+    .then((response) => response.text())
+    .then((result) => console.info(result))
+    .catch((error) => console.error(error));
+  return json({ success: true });
+};
